Fold MultiFileReadAhead into ScanReadAhead
